Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
pyjwt>=2.9.0
python-multipart>=0.0.12
openai>=1.55.0
pydantic>=2.0.0
pydantic-settings>=2.3.0
httpx>=0.27.0
dnspython==2.8.0
email-validator==2.3.0
Pillow>=10.0.0
bcrypt>=4.0.0
groq>=0.9.0
pytesseract>=0.3.10
pymupdf>=1.24.0
openpyxl>=3.1.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
locust>=2.32.0